pyproject.toml
[tool.poetry]
name = "blacksheep-clean-architecture"
version = "1.0beta"
description = ""
authors = ["Gleb Garanin <glebgar567@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.26"
alembic = "^1.7.5"
asyncpg = "^0.25.0"
dynaconf = { version="^3.1.7", extras=["all"] }
blacksheep = "^1.2.2"
uvloop = "^0.16.0"
gunicorn = "^20.1.0"
uvicorn = "^0.16.0"
pydantic = "^1.8.2"
orjson = "^3.6.5"
colorlog = "6.6.0"
guardpost = "^0.0.9"
PyJWT = "^2.3.0"
py-bcrypt = "^0.4"
blacksheep-prometheus = "^0.1.6"
redis = "^4.0.2"
apscheduler-di = "^0.0.2"
aio-pika = "^6.8.0"
structlog = "^21.4.0"
argon2-cffi = "^21.3.0"
structlog-sentry = "^1.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
mypy = "^0.910"
sqlalchemy2-stubs = "^0.0.2-alpha.19"
black = "^21.12b0"
types-ujson = "^4.2.0"
isort = { version="^5.10.0", extras=["requirements_deprecated_finder","pipfile_deprecated_finder"]}
devtools = "^0.8.0"
pytest-lazy-fixture = "^0.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
